🚚 change version code data type to int from long

+++ src/main/kotlin/brut/androlib/meta/MetaInfo.kt
@@ -13,7 +13,7 @@ data class MetaInfo(
     )
 
     data class VersionInfo(
-        var versionCode: Long? = null,
+        var versionCode: Int? = null,
         var versionName: String? = null
     )
 }
