+++ src/main/kotlin/com/theapache64/stackzy/data/repo/ResultRepo.kt
@@ -12,7 +12,7 @@ class ResultRepo @Inject constructor(
     private val moshi: Moshi
 ) {
     fun add(result: Result) = apiInterface.addResult(result)
-    fun findResult(packageName: String, versionCode: Long) = apiInterface.getResult(packageName, versionCode)
+    fun findResult(packageName: String, versionCode: Int) = apiInterface.getResult(packageName, versionCode)
 
     private val gradleInfoAdapter by lazy {
         GradleInfoJsonAdapter(moshi)
